+++ NEWS.md
@@ -1,3 +1,4 @@
 # nethist 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+-   Initial release of nethist R package
+